requirements.txt
rasterio==1.3.9
opencv-python==4.8.0.76
numpy==1.26.1
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
pandas==2.0.3
scikit-learn==1.2.2
streamlit==1.25.0
pillow==10.0.0